Refactor player into modules
