+++ app/components/player/components/TrackInfo.tsx
@@ -1,0 +1,41 @@
+import React from 'react';
+
+interface Props {
+  cover: string;
+  title: string;
+  artist: string;
+  theme: 'light' | 'dark';
+}
+
+export default function TrackInfo({ cover, title, artist, theme }: Props) {
+  return (
+    <div className="flex items-center gap-3 min-w-0 sm:min-w-[220px] flex-shrink-0">
+      <img
+        src={cover}
+        alt="cover"
+        className="h-12 w-12 rounded-full shadow-sm object-cover"
+        onError={(e) => {
+          e.currentTarget.src =
+            "data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='96' height='96'><rect width='100%' height='100%' rx='12' ry='12' fill='%23e5e7eb'/><text x='50%' y='52%' dominant-baseline='middle' text-anchor='middle' font-family='Inter, system-ui, sans-serif' font-size='12' fill='%239ca3af'>No cover</text></svg>";
+        }}
+      />
+      <div className="min-w-0 hidden sm:block">
+        <div
+          className={`text-sm font-semibold tracking-[-0.01em] truncate ${
+            theme === 'dark' ? 'text-slate-200' : 'text-slate-800'
+          }`}
+          aria-label="current track title"
+        >
+          {title}
+        </div>
+        <div
+          className={`text-xs -mt-0.5 truncate ${
+            theme === 'dark' ? 'text-slate-400' : 'text-slate-500'
+          }`}
+        >
+          {artist}
+        </div>
+      </div>
+    </div>
+  );
+}
